userspace/nat: split destination-NAT compiler into nat_destination.go
Continues the #4421 (codex-173 #5) pure code-motion split. Moves the
destination-NAT declarations verbatim out of nat.go into
nat_destination.go: dnatDestinationParts (host-vs-prefix classifier),
dnatPoolHostIP (pool single-host validator), buildDestinationNATSnapshots
and its feed-aware buildDestinationNATSnapshotsWithFeeds worker.

This is the largest single family in the file. No body edits, no
renames. The DNAT builder still calls the cross-type shared helpers that
remain in nat.go (natCounterID, appendNATSource/DestinationAddressName,
coalescePortRanges, appPortsFromSpec, natNeverMatchPortRange); the two
dnat* helpers are DNAT-local and move with it.

Validation: go build ./... green.
